+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### master
+- restore pane processes
 
 ### v0.0.4, 2014-08-26
 - restore pane layout for each window